requirements.txt
pinecone-client
langchain
openai
PyPDF2
streamlit
sentence-transformers
pinecone-text
pinecone-notebooks
python-dotenv
langchain-community
langchain-huggingface
